share GraphQL Soup filter materialization
