+++ crates/document_sub_type/Cargo.toml
@@ -4,11 +4,15 @@ name = "document_sub_type"
 publish = false
 version = "0.1.0"
 
+[features]
+default = ["schema", "sqlx"]
+schema = ["dep:schemars", "dep:utoipa"]
+sqlx = ["dep:sqlx"]
+
 [dependencies]
 serde = { workspace = true }
-sqlx = { workspace = true }
+sqlx = { workspace = true, optional = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
-utoipa = { workspace = true }
-schemars = {workspace = true}
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
+utoipa = { workspace = true, optional = true }
+schemars = { workspace = true, optional = true }
