+++ crates/document_sub_type/src/lib.rs
@@ -1,5 +1,4 @@
 //! This crate contains the document sub type enum and various logic for it
-use schemars::JsonSchema;
 
 /// The document sub type enum represents all values of document sub types.
 /// These values should match the `document_sub_type_value` table in macrodb.
@@ -9,17 +8,19 @@ use schemars::JsonSchema;
     Eq,
     PartialEq,
     Debug,
-    utoipa::ToSchema,
     Clone,
     Copy,
-    sqlx::Type,
     strum::EnumString,
     strum::Display,
-    JsonSchema,
 )]
+#[cfg_attr(feature = "schema", derive(utoipa::ToSchema, schemars::JsonSchema))]
+#[cfg_attr(feature = "sqlx", derive(sqlx::Type))]
 #[serde(rename_all = "lowercase")]
 #[strum(serialize_all = "snake_case")]
-#[sqlx(type_name = "\"document_sub_type_value\"", rename_all = "lowercase")]
+#[cfg_attr(
+    feature = "sqlx",
+    sqlx(type_name = "\"document_sub_type_value\"", rename_all = "lowercase")
+)]
 pub enum DocumentSubType {
     /// A task document
     Task,
