+++ crates/filter_ast/Cargo.toml
@@ -7,4 +7,3 @@ version = "0.1.0"
 [dependencies]
 recursion = { workspace = true }
 serde = { workspace = true }
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
