+++ crates/graphql_soup/Cargo.toml
@@ -22,6 +22,7 @@ entity_access = { path = "../entity_access", default-features = false, features 
 ] }
 filter_ast = { path = "../filter_ast" }
 graphql_common = { path = "../graphql_common" }
+graphql_soup_filter_input = { path = "../graphql_soup_filter_input" }
 graphql_email = { path = "../graphql_email" }
 graphql_permission = { path = "../graphql_permission" }
 item_filters = { path = "../item_filters" }
