+++ crates/graphql_soup/src/inputs.rs
@@ -1,33 +1,8 @@
-use std::str::FromStr;
-
 use async_graphql::{Enum, ID};
-use chrono::{DateTime, Utc};
-use document_sub_type::DocumentSubType;
-use filter_ast::Expr;
-use graphql_common::{
-    GraphqlPropertiesExpr, GraphqlPropertyEntityType, IntoFilterExpr, filter_expr_input,
-    optional_tree, parse_id, parse_macro_user_id,
-};
-use item_filters::{
-    CallStatus, SharedEmailFilter,
-    ast::{
-        CrmScope, EmailFilterAst, EntityFilterAst,
-        calendar_event::CalendarEventLiteral,
-        call::CallLiteral,
-        channel::{ChannelLiteral, ChannelThreadLiteral, ChannelTypeFilter},
-        chat::{ChatLiteral, ChatRole},
-        crm_company::CrmCompanyLiteral,
-        date::DateLiteral,
-        document::DocumentLiteral,
-        email::{Email, EmailLiteral},
-        foreign_entity::ForeignEntityLiteral,
-        project::ProjectLiteral,
-        properties::PropertyEntityType,
-        reminder::ReminderLiteral,
-    },
-};
-use macro_user_id::{cowlike::CowLike, email::EmailStr, user_id::MacroUserIdStr};
-use model_file_type::FileType;
+use graphql_common::{GraphqlPropertyEntityType, parse_id};
+use graphql_soup_filter_input::GraphqlEntityFilterAst;
+use item_filters::ast::{EntityFilterAst, properties::PropertyEntityType};
+use macro_user_id::user_id::MacroUserIdStr;
 use models_grouping::{GroupByField, GroupingConfig};
 use models_pagination::{Base64Str, CursorWithValAndFilter, Query, SimpleSortMethod};
 use soup::domain::models::{
@@ -402,820 +377,8 @@ impl GraphqlEmailView {
     }
 }
 
-/// GraphQL input mirroring `item_filters::ast::EntityFilterAst`.
-#[derive(async_graphql::InputObject)]
-struct GraphqlEntityFilterAst {
-    /// The calendar event filter to apply.
-    calendar_event_filter: Option<GraphqlCalendarEventExpr>,
-    /// The document filter to apply.
-    document_filter: Option<GraphqlDocumentExpr>,
-    /// The project filter to apply.
-    project_filter: Option<GraphqlProjectExpr>,
-    /// The chat filter to apply.
-    chat_filter: Option<GraphqlChatExpr>,
-    /// The email filter to apply.
-    email_filter: Option<GraphqlEmailFilterAst>,
-    /// The channel filter to apply.
-    channel_filter: Option<GraphqlChannelExpr>,
-    /// The channel thread filter to apply.
-    channel_thread_filter: Option<GraphqlChannelThreadExpr>,
-    /// The call filter to apply.
-    call_filter: Option<GraphqlCallExpr>,
-    /// The crm company filter to apply.
-    crm_company_filter: Option<GraphqlCrmCompanyExpr>,
-    /// The foreign entity filter to apply.
-    foreign_entity_filter: Option<GraphqlForeignEntityExpr>,
-    /// The reminder filter to apply.
-    reminder_filter: Option<GraphqlReminderExpr>,
-    /// The properties filter to apply.
-    properties_filter: Option<GraphqlPropertiesExpr>,
-}
-
-impl GraphqlEntityFilterAst {
-    /// Convert this value into the ast representation.
-    fn into_ast(self) -> async_graphql::Result<EntityFilterAst> {
-        Ok(EntityFilterAst {
-            calendar_event_filter: optional_tree(self.calendar_event_filter)?,
-            document_filter: optional_tree(self.document_filter)?,
-            project_filter: optional_tree(self.project_filter)?,
-            chat_filter: optional_tree(self.chat_filter)?,
-            email_filter: self
-                .email_filter
-                .map(GraphqlEmailFilterAst::into_ast)
-                .transpose()?
-                .unwrap_or_default(),
-            channel_filter: optional_tree(self.channel_filter)?,
-            channel_thread_filter: optional_tree(self.channel_thread_filter)?,
-            call_filter: optional_tree(self.call_filter)?,
-            crm_company_filter: optional_tree(self.crm_company_filter)?,
-            foreign_entity_filter: optional_tree(self.foreign_entity_filter)?,
-            reminder_filter: optional_tree(self.reminder_filter)?,
-            properties_filter: optional_tree(self.properties_filter)?,
-        })
-    }
-}
-
-filter_expr_input!(
-    GraphqlCalendarEventExpr,
-    GraphqlCalendarEventBinaryExpr,
-    GraphqlCalendarEventLiteral,
-    CalendarEventLiteral,
-    "CalendarEventFilterExpr"
-);
-filter_expr_input!(
-    GraphqlDocumentExpr,
-    GraphqlDocumentBinaryExpr,
-    GraphqlDocumentLiteral,
-    DocumentLiteral,
-    "DocumentFilterExpr"
-);
-
-/// GraphQL input representing a calendar event literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlCalendarEventLiteral {
-    /// Canonical event id.
-    Id(ID),
-    /// Event status.
-    Status(String),
-    /// Master start must be before this RFC3339 instant.
-    StartsBefore(String),
-    /// Master end must be after this RFC3339 instant.
-    EndsAfter(String),
-    /// Attendee email.
-    Attendee(String),
-    /// Organizer email.
-    Organizer(String),
-    /// Notification done state for the requester.
-    NotificationDone(bool),
-    /// Notification seen state for the requester.
-    NotificationSeen(bool),
-}
-
-impl IntoFilterExpr<CalendarEventLiteral> for GraphqlCalendarEventLiteral {
-    fn into_expr(self) -> async_graphql::Result<Expr<CalendarEventLiteral>> {
-        let parse_date = |value: String| {
-            DateTime::parse_from_rfc3339(&value)
-                .map(|date| date.with_timezone(&Utc))
-                .map_err(|error| {
-                    async_graphql::Error::new(format!(
-                        "invalid RFC3339 calendar date `{value}`: {error}"
-                    ))
-                })
-        };
-        Ok(Expr::val(match self {
-            Self::Id(id) => CalendarEventLiteral::Id(parse_id(id, "id")?),
-            Self::Status(status) => CalendarEventLiteral::Status(status.to_ascii_lowercase()),
-            Self::StartsBefore(value) => CalendarEventLiteral::StartsBefore(parse_date(value)?),
-            Self::EndsAfter(value) => CalendarEventLiteral::EndsAfter(parse_date(value)?),
-            Self::Attendee(email) => CalendarEventLiteral::Attendee(email.to_ascii_lowercase()),
-            Self::Organizer(email) => CalendarEventLiteral::Organizer(email.to_ascii_lowercase()),
-            Self::NotificationDone(done) => CalendarEventLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => CalendarEventLiteral::NotificationSeen(seen),
-        }))
-    }
-}
-filter_expr_input!(
-    GraphqlProjectExpr,
-    GraphqlProjectBinaryExpr,
-    GraphqlProjectLiteral,
-    ProjectLiteral,
-    "ProjectFilterExpr"
-);
-filter_expr_input!(
-    GraphqlChatExpr,
-    GraphqlChatBinaryExpr,
-    GraphqlChatLiteral,
-    ChatLiteral,
-    "ChatFilterExpr"
-);
-filter_expr_input!(
-    GraphqlEmailExpr,
-    GraphqlEmailBinaryExpr,
-    GraphqlEmailLiteral,
-    EmailLiteral,
-    "EmailFilterExpr"
-);
-filter_expr_input!(
-    GraphqlChannelExpr,
-    GraphqlChannelBinaryExpr,
-    GraphqlChannelLiteral,
-    ChannelLiteral,
-    "ChannelFilterExpr"
-);
-filter_expr_input!(
-    GraphqlChannelThreadExpr,
-    GraphqlChannelThreadBinaryExpr,
-    GraphqlChannelThreadLiteral,
-    ChannelThreadLiteral,
-    "ChannelThreadFilterExpr"
-);
-filter_expr_input!(
-    GraphqlCallExpr,
-    GraphqlCallBinaryExpr,
-    GraphqlCallLiteral,
-    CallLiteral,
-    "CallFilterExpr"
-);
-filter_expr_input!(
-    GraphqlCrmCompanyExpr,
-    GraphqlCrmCompanyBinaryExpr,
-    GraphqlCrmCompanyLiteral,
-    CrmCompanyLiteral,
-    "CrmCompanyFilterExpr"
-);
-filter_expr_input!(
-    GraphqlForeignEntityExpr,
-    GraphqlForeignEntityBinaryExpr,
-    GraphqlForeignEntityLiteral,
-    ForeignEntityLiteral,
-    "ForeignEntityFilterExpr"
-);
-filter_expr_input!(
-    GraphqlReminderExpr,
-    GraphqlReminderBinaryExpr,
-    GraphqlReminderLiteral,
-    ReminderLiteral,
-    "ReminderFilterExpr"
-);
-/// GraphQL input representing the email filter ast.
-#[derive(async_graphql::InputObject)]
-struct GraphqlEmailFilterAst {
-    /// The tree.
-    tree: Option<GraphqlEmailExpr>,
-    /// The crm scope.
-    crm_scope: Option<GraphqlCrmScope>,
-}
-
-impl GraphqlEmailFilterAst {
-    /// Convert this value into the ast representation.
-    fn into_ast(self) -> async_graphql::Result<EmailFilterAst> {
-        Ok(EmailFilterAst {
-            tree: optional_tree(self.tree)?,
-            crm_scope: self.crm_scope.map(GraphqlCrmScope::into_ast).transpose()?,
-        })
-    }
-}
-
-/// GraphQL input representing the crm scope.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlCrmScope {
-    /// The domains option.
-    Domains(Vec<String>),
-    /// The addresses option.
-    Addresses(Vec<String>),
-}
-
-impl GraphqlCrmScope {
-    /// Convert this value into the ast representation.
-    fn into_ast(self) -> async_graphql::Result<CrmScope> {
-        match self {
-            Self::Domains(domains) if domains.is_empty() => Err(async_graphql::Error::new(
-                "CrmScope.domains cannot be empty",
-            )),
-            Self::Domains(domains) => Ok(CrmScope::Domains(domains)),
-            Self::Addresses(addresses) if addresses.is_empty() => Err(async_graphql::Error::new(
-                "CrmScope.addresses cannot be empty",
-            )),
-            Self::Addresses(addresses) => Ok(CrmScope::Addresses(addresses)),
-        }
-    }
-}
-
-/// GraphQL input representing the date literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlDateLiteral {
-    /// The gt option.
-    Gt(String),
-    /// The lt option.
-    Lt(String),
-    /// The gte option.
-    Gte(String),
-    /// The lte option.
-    Lte(String),
-}
-
-impl GraphqlDateLiteral {
-    /// Parse an email address from a GraphQL string value.
-    fn parse(value: String) -> async_graphql::Result<DateTime<Utc>> {
-        DateTime::parse_from_rfc3339(&value)
-            .map(|dt| dt.with_timezone(&Utc))
-            .map_err(|err| {
-                async_graphql::Error::new(format!("invalid RFC3339 date `{value}`: {err}"))
-            })
-    }
-
-    /// Convert this value into the ast representation.
-    fn into_ast(self) -> async_graphql::Result<DateLiteral> {
-        Ok(match self {
-            Self::Gt(value) => DateLiteral::GreaterThan(Self::parse(value)?),
-            Self::Lt(value) => DateLiteral::LessThan(Self::parse(value)?),
-            Self::Gte(value) => DateLiteral::GreaterThanOrEqual(Self::parse(value)?),
-            Self::Lte(value) => DateLiteral::LessThanOrEqual(Self::parse(value)?),
-        })
-    }
-}
-
-/// GraphQL input representing the document literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlDocumentLiteral {
-    /// The file type option.
-    FileType(String),
-    /// The id option.
-    Id(ID),
-    /// The project id option.
-    ProjectId(ID),
-    /// The owner option.
-    Owner(String),
-    /// The importance option.
-    Importance(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-    /// The include cbm atm nc option.
-    IncludeCbmAtmNc(bool),
-    /// The sub type option.
-    SubType(GraphqlDocumentSubType),
-    /// The file assoc option.
-    FileAssoc(String),
-    /// The is email attachment option.
-    IsEmailAttachment(bool),
-    /// The created at option.
-    CreatedAt(GraphqlDateLiteral),
-    /// The updated at option.
-    UpdatedAt(GraphqlDateLiteral),
-}
-
-impl IntoFilterExpr<DocumentLiteral> for GraphqlDocumentLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<DocumentLiteral>> {
-        let literal = match self {
-            Self::FileAssoc(value) => {
-                let (_, file_types) = item_filters::ast::document::parse_to_file_types(&value)
-                    .map_err(|err| async_graphql::Error::new(err.to_string()))?;
-                return file_types
-                    .map(|file_type| Expr::val(DocumentLiteral::FileType(file_type)))
-                    .reduce(Expr::or)
-                    .ok_or_else(|| {
-                        async_graphql::Error::new("fileAssoc expansion cannot be empty")
-                    });
-            }
-            Self::FileType(value) => {
-                DocumentLiteral::FileType(FileType::from_str(&value).map_err(|err| {
-                    async_graphql::Error::new(format!("invalid fileType `{value}`: {err}"))
-                })?)
-            }
-            Self::Id(id) => DocumentLiteral::Id(parse_id(id, "id")?),
-            Self::ProjectId(id) => DocumentLiteral::ProjectId(parse_id(id, "projectId")?),
-            Self::Owner(owner) => DocumentLiteral::Owner(parse_macro_user_id(owner, "owner")?),
-            Self::Importance(importance) => DocumentLiteral::Importance(importance),
-            Self::NotificationDone(done) => DocumentLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => DocumentLiteral::NotificationSeen(seen),
-            Self::IncludeCbmAtmNc(include) => DocumentLiteral::IncludeCbmAtmNc(include),
-            Self::SubType(sub_type) => DocumentLiteral::SubType(sub_type.into_model()),
-            Self::IsEmailAttachment(value) => DocumentLiteral::IsEmailAttachment(value),
-            Self::CreatedAt(date) => DocumentLiteral::CreatedAt(date.into_ast()?),
-            Self::UpdatedAt(date) => DocumentLiteral::UpdatedAt(date.into_ast()?),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the document sub type.
-#[derive(Enum, Copy, Clone, Eq, PartialEq)]
-enum GraphqlDocumentSubType {
-    /// The task option.
-    Task,
-    /// The snippet option.
-    Snippet,
-    /// The skill option.
-    Skill,
-}
-
-impl GraphqlDocumentSubType {
-    /// Convert this GraphQL subtype into the document-filter model.
-    fn into_model(self) -> DocumentSubType {
-        match self {
-            Self::Task => DocumentSubType::Task,
-            Self::Snippet => DocumentSubType::Snippet,
-            Self::Skill => DocumentSubType::Skill,
-        }
-    }
-}
-
-/// GraphQL input representing the project literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlProjectLiteral {
-    /// The project id option.
-    ProjectId(ID),
-    /// The project id self option.
-    ProjectIdSelf(ID),
-    /// The owner option.
-    Owner(String),
-    /// The importance option.
-    Importance(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-    /// The created at option.
-    CreatedAt(GraphqlDateLiteral),
-    /// The updated at option.
-    UpdatedAt(GraphqlDateLiteral),
-}
-
-impl IntoFilterExpr<ProjectLiteral> for GraphqlProjectLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ProjectLiteral>> {
-        let literal = match self {
-            Self::ProjectId(id) => ProjectLiteral::ProjectId(parse_id(id, "projectId")?),
-            Self::ProjectIdSelf(id) => {
-                ProjectLiteral::ProjectIdSelf(parse_id(id, "projectIdSelf")?)
-            }
-            Self::Owner(owner) => ProjectLiteral::Owner(parse_macro_user_id(owner, "owner")?),
-            Self::Importance(importance) => ProjectLiteral::Importance(importance),
-            Self::NotificationDone(done) => ProjectLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => ProjectLiteral::NotificationSeen(seen),
-            Self::CreatedAt(date) => ProjectLiteral::CreatedAt(date.into_ast()?),
-            Self::UpdatedAt(date) => ProjectLiteral::UpdatedAt(date.into_ast()?),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the chat literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlChatLiteral {
-    /// The project id option.
-    ProjectId(ID),
-    /// The role option.
-    Role(GraphqlChatRole),
-    /// The chat id option.
-    ChatId(ID),
-    /// The owner option.
-    Owner(String),
-    /// The importance option.
-    Importance(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-    /// The created at option.
-    CreatedAt(GraphqlDateLiteral),
-    /// The updated at option.
-    UpdatedAt(GraphqlDateLiteral),
-}
-
-impl IntoFilterExpr<ChatLiteral> for GraphqlChatLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ChatLiteral>> {
-        let literal = match self {
-            Self::ProjectId(id) => ChatLiteral::ProjectId(parse_id(id, "projectId")?),
-            Self::Role(role) => ChatLiteral::Role(role.into_model()),
-            Self::ChatId(id) => ChatLiteral::ChatId(parse_id(id, "chatId")?),
-            Self::Owner(owner) => ChatLiteral::Owner(parse_macro_user_id(owner, "owner")?),
-            Self::Importance(importance) => ChatLiteral::Importance(importance),
-            Self::NotificationDone(done) => ChatLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => ChatLiteral::NotificationSeen(seen),
-            Self::CreatedAt(date) => ChatLiteral::CreatedAt(date.into_ast()?),
-            Self::UpdatedAt(date) => ChatLiteral::UpdatedAt(date.into_ast()?),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the chat role.
-#[derive(Enum, Copy, Clone, Eq, PartialEq)]
-enum GraphqlChatRole {
-    /// The user option.
-    User,
-    /// The system option.
-    System,
-    /// The assistant option.
-    Assistant,
-}
-
-impl GraphqlChatRole {
-    /// Convert this GraphQL role into the chat-filter model.
-    fn into_model(self) -> ChatRole {
-        match self {
-            Self::User => ChatRole::User,
-            Self::System => ChatRole::System,
-            Self::Assistant => ChatRole::Assistant,
-        }
-    }
-}
-
-/// GraphQL input representing the email literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlEmailLiteral {
-    /// The sender option.
-    Sender(GraphqlEmailValue),
-    /// The cc option.
-    Cc(GraphqlEmailValue),
-    /// The bcc option.
-    Bcc(GraphqlEmailValue),
-    /// The recipient option.
-    Recipient(GraphqlEmailValue),
-    /// The thread id option.
-    ThreadId(ID),
-    /// The owner option.
-    Owner(ID),
-    /// The project id option.
-    ProjectId(String),
-    /// The importance option.
-    Importance(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-    /// The shared option.
-    Shared(GraphqlSharedEmailFilter),
-    /// The calendar only option.
-    CalendarOnly(bool),
-    /// The created at option.
-    CreatedAt(GraphqlDateLiteral),
-    /// The updated at option.
-    UpdatedAt(GraphqlDateLiteral),
-}
-
-impl IntoFilterExpr<EmailLiteral> for GraphqlEmailLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<EmailLiteral>> {
-        let literal = match self {
-            Self::Sender(value) => EmailLiteral::Sender(value.into_ast()?),
-            Self::Cc(value) => EmailLiteral::Cc(value.into_ast()?),
-            Self::Bcc(value) => EmailLiteral::Bcc(value.into_ast()?),
-            Self::Recipient(value) => EmailLiteral::Recipient(value.into_ast()?),
-            Self::ThreadId(id) => EmailLiteral::ThreadId(parse_id(id, "threadId")?),
-            Self::Owner(id) => EmailLiteral::Owner(parse_id(id, "owner")?),
-            Self::ProjectId(id) => EmailLiteral::ProjectId(id),
-            Self::Importance(importance) => EmailLiteral::Importance(importance),
-            Self::NotificationDone(done) => EmailLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => EmailLiteral::NotificationSeen(seen),
-            Self::Shared(shared) => EmailLiteral::Shared(shared.into_model()),
-            Self::CalendarOnly(calendar_only) => EmailLiteral::CalendarOnly(calendar_only),
-            Self::CreatedAt(date) => EmailLiteral::CreatedAt(date.into_ast()?),
-            Self::UpdatedAt(date) => EmailLiteral::UpdatedAt(date.into_ast()?),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the email value.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlEmailValue {
-    /// The partial option.
-    Partial(String),
-    /// The complete option.
-    Complete(String),
-    /// The domain option.
-    Domain(String),
-}
-
-impl GraphqlEmailValue {
-    /// Convert this value into the ast representation.
-    fn into_ast(self) -> async_graphql::Result<Email> {
-        Ok(match self {
-            Self::Partial(value) => Email::Partial(value),
-            Self::Complete(value) => Email::Complete(
-                EmailStr::parse_from_str(&value)
-                    .map(CowLike::into_owned)
-                    .map_err(|err| {
-                        async_graphql::Error::new(format!(
-                            "invalid complete email `{value}`: {err}"
-                        ))
-                    })?,
-            ),
-            Self::Domain(value) => Email::Domain(value),
-        })
-    }
-}
-
-/// GraphQL input representing the shared email filter.
-#[derive(Enum, Copy, Clone, Eq, PartialEq)]
-enum GraphqlSharedEmailFilter {
-    /// The exclude option.
-    Exclude,
-    /// The include option.
-    Include,
-    /// The only option.
-    Only,
-}
-
-impl GraphqlSharedEmailFilter {
-    /// Convert this GraphQL option into the shared-email filter model.
-    fn into_model(self) -> SharedEmailFilter {
-        match self {
-            Self::Exclude => SharedEmailFilter::Exclude,
-            Self::Include => SharedEmailFilter::Include,
-            Self::Only => SharedEmailFilter::Only,
-        }
-    }
-}
-
-/// GraphQL input representing the channel literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlChannelLiteral {
-    /// The thread id option.
-    ThreadId(ID),
-    /// The mention option.
-    Mention(String),
-    /// The organization id option.
-    OrganizationId(i64),
-    /// The team id option.
-    TeamId(ID),
-    /// The channel id option.
-    ChannelId(ID),
-    /// The sender option.
-    Sender(String),
-    /// The channel type option.
-    ChannelType(GraphqlChannelTypeFilter),
-    /// The importance option.
-    Importance(bool),
-    /// The is participant option. Filters by whether the requesting user is an
-    /// active participant; its presence widens the candidate set to team channels
-    /// of the user's teams they have not joined.
-    IsParticipant(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-}
-
-impl IntoFilterExpr<ChannelLiteral> for GraphqlChannelLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ChannelLiteral>> {
-        let literal = match self {
-            Self::ThreadId(id) => ChannelLiteral::ThreadId(parse_id(id, "threadId")?),
-            Self::Mention(mention) => {
-                ChannelLiteral::Mention(parse_macro_user_id(mention, "mention")?)
-            }
-            Self::OrganizationId(id) => ChannelLiteral::OrganizationId(id),
-            Self::TeamId(id) => ChannelLiteral::TeamId(parse_id(id, "teamId")?),
-            Self::ChannelId(id) => ChannelLiteral::ChannelId(parse_id(id, "channelId")?),
-            Self::Sender(sender) => ChannelLiteral::Sender(parse_macro_user_id(sender, "sender")?),
-            Self::ChannelType(channel_type) => {
-                ChannelLiteral::ChannelType(channel_type.into_model())
-            }
-            Self::Importance(importance) => ChannelLiteral::Importance(importance),
-            Self::IsParticipant(is_participant) => ChannelLiteral::IsParticipant(is_participant),
-            Self::NotificationDone(done) => ChannelLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => ChannelLiteral::NotificationSeen(seen),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the channel type filter.
-#[derive(Enum, Copy, Clone, Eq, PartialEq)]
-enum GraphqlChannelTypeFilter {
-    /// The public option.
-    Public,
-    /// The private option.
-    Private,
-    /// The direct message option.
-    DirectMessage,
-    /// The team option.
-    Team,
-}
-
-impl GraphqlChannelTypeFilter {
-    /// Convert this GraphQL option into the channel-filter model.
-    fn into_model(self) -> ChannelTypeFilter {
-        match self {
-            Self::Public => ChannelTypeFilter::Public,
-            Self::Private => ChannelTypeFilter::Private,
-            Self::DirectMessage => ChannelTypeFilter::DirectMessage,
-            Self::Team => ChannelTypeFilter::Team,
-        }
-    }
-}
-
-/// GraphQL input representing the channel thread literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlChannelThreadLiteral {
-    /// The thread id option.
-    ThreadId(ID),
-    /// The channel id option.
-    ChannelId(ID),
-    /// The root sender option.
-    RootSender(String),
-    /// The participant option.
-    Participant(String),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-}
-
-impl IntoFilterExpr<ChannelThreadLiteral> for GraphqlChannelThreadLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ChannelThreadLiteral>> {
-        let literal = match self {
-            Self::ThreadId(id) => ChannelThreadLiteral::ThreadId(parse_id(id, "threadId")?),
-            Self::ChannelId(id) => ChannelThreadLiteral::ChannelId(parse_id(id, "channelId")?),
-            Self::RootSender(sender) => {
-                ChannelThreadLiteral::RootSender(parse_macro_user_id(sender, "rootSender")?)
-            }
-            Self::Participant(participant) => {
-                ChannelThreadLiteral::Participant(parse_macro_user_id(participant, "participant")?)
-            }
-            Self::NotificationDone(done) => ChannelThreadLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => ChannelThreadLiteral::NotificationSeen(seen),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the call literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlCallLiteral {
-    /// The call id option.
-    CallId(ID),
-    /// The channel id option.
-    ChannelId(ID),
-    /// The speaker option.
-    Speaker(String),
-    /// The status option.
-    Status(GraphqlCallStatus),
-    /// The attended option.
-    Attended(bool),
-}
-
-impl IntoFilterExpr<CallLiteral> for GraphqlCallLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<CallLiteral>> {
-        let literal = match self {
-            Self::CallId(id) => CallLiteral::CallId(parse_id(id, "callId")?),
-            Self::ChannelId(id) => CallLiteral::ChannelId(parse_id(id, "channelId")?),
-            Self::Speaker(speaker) => {
-                CallLiteral::Speaker(parse_macro_user_id(speaker, "speaker")?)
-            }
-            Self::Status(status) => CallLiteral::Status(status.into_model()),
-            Self::Attended(attended) => CallLiteral::Attended(attended),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the call status.
-#[derive(Enum, Copy, Clone, Eq, PartialEq)]
-enum GraphqlCallStatus {
-    /// The attended option.
-    Attended,
-    /// The missed option.
-    Missed,
-    /// The unattended option.
-    Unattended,
-}
-
-impl GraphqlCallStatus {
-    /// Convert this GraphQL status into the call-filter model.
-    fn into_model(self) -> CallStatus {
-        match self {
-            Self::Attended => CallStatus::Attended,
-            Self::Missed => CallStatus::Missed,
-            Self::Unattended => CallStatus::Unattended,
-        }
-    }
-}
-
-/// GraphQL input representing the reminder literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlReminderLiteral {
-    /// Opt this query into reminders at all. Reminders are off by default, so
-    /// without this (or an `id`/`entity`) Soup omits them entirely — a filter
-    /// of only `completed` would otherwise silently match nothing. Must be
-    /// `true`; there is no literal for excluding reminders, that is the default.
-    Include(bool),
-    /// The id option.
-    Id(ID),
-    /// The referenced entity, as `"{type}:{id}"`.
-    Entity(String),
-    /// Whether the owner has marked the reminder done.
-    Completed(bool),
-    /// Whether the reminder has come due and is awaiting its owner.
-    Fired(bool),
-}
-
-impl IntoFilterExpr<ReminderLiteral> for GraphqlReminderLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ReminderLiteral>> {
-        let literal = match self {
-            // `include: false` is the default, not a literal — accepting it
-            // would opt the query in, the opposite of what was asked.
-            Self::Include(false) => {
-                return Err(async_graphql::Error::new(
-                    "reminder `include` must be true; omit the filter to exclude reminders",
-                ));
-            }
-            Self::Include(true) => ReminderLiteral::Include,
-            Self::Id(id) => ReminderLiteral::Id(parse_id(id, "id")?),
-            Self::Entity(entity) => ReminderLiteral::Entity(entity),
-            Self::Completed(completed) => ReminderLiteral::Completed(completed),
-            Self::Fired(fired) => ReminderLiteral::Fired(fired),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the crm company literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlCrmCompanyLiteral {
-    /// The id option.
-    Id(ID),
-    /// The hidden option.
-    Hidden(bool),
-}
-
-impl IntoFilterExpr<CrmCompanyLiteral> for GraphqlCrmCompanyLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<CrmCompanyLiteral>> {
-        let literal = match self {
-            Self::Id(id) => CrmCompanyLiteral::Id(parse_id(id, "id")?),
-            Self::Hidden(hidden) => CrmCompanyLiteral::Hidden(hidden),
-        };
-        Ok(Expr::val(literal))
-    }
-}
-
-/// GraphQL input representing the foreign entity literal.
-#[derive(async_graphql::OneofObject)]
-enum GraphqlForeignEntityLiteral {
-    /// The id option.
-    Id(ID),
-    /// The foreign entity id option.
-    ForeignEntityId(String),
-    /// The foreign entity source option.
-    ForeignEntitySource(String),
-    /// The includes me option.
-    IncludesMe(bool),
-    /// The notification done option.
-    NotificationDone(bool),
-    /// The notification seen option.
-    NotificationSeen(bool),
-}
-
-impl IntoFilterExpr<ForeignEntityLiteral> for GraphqlForeignEntityLiteral {
-    /// Convert this value into the expr representation.
-    fn into_expr(self) -> async_graphql::Result<Expr<ForeignEntityLiteral>> {
-        let literal = match self {
-            Self::Id(id) => ForeignEntityLiteral::Id(parse_id(id, "id")?),
-            Self::ForeignEntityId(id) => ForeignEntityLiteral::ForeignEntityId(id),
-            Self::ForeignEntitySource(source) => ForeignEntityLiteral::ForeignEntitySource(source),
-            Self::IncludesMe(true) => ForeignEntityLiteral::IncludesMe,
-            Self::IncludesMe(false) => {
-                return Err(async_graphql::Error::new(
-                    "ForeignEntityLiteral.includesMe must be true",
-                ));
-            }
-            Self::NotificationDone(done) => ForeignEntityLiteral::NotificationDone(done),
-            Self::NotificationSeen(seen) => ForeignEntityLiteral::NotificationSeen(seen),
-        };
-        Ok(Expr::val(literal))
-    }
-}
+// Filter input types and their authoritative materialization live in
+// `graphql_soup_filter_input`, shared with the browser cache composition root.
 
 /// GraphQL representation of supported simple Soup sorts.
 #[derive(Enum, Copy, Clone, Eq, PartialEq)]
