+++ crates/model_file_type/Cargo.toml
@@ -17,4 +17,3 @@ serde_json = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
 utoipa = { workspace = true, optional = true }
-workspace-hack = { version = "0.1", path = "../workspace-hack" }
